Put classes into packages
