+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,6 @@
-public class DukeException extends Exception {
+package duke;
+
+class DukeException extends Exception {
 
     protected DukeException(String message) {
         super(message);
