+++ src/main/java/duke/Ui.java
@@ -1,4 +1,8 @@
-public class Ui {
+package duke;
+
+import duke.command.Commands;
+
+class Ui {
     public void outputMessage(Commands command) {
         String message = command.getCommand();
         System.out.println(String.format("%4s%s", " ", message));
