+++ src/main/java/duke/task/Event.java
@@ -1,7 +1,9 @@
+package duke.task;
+
 import java.time.LocalDate;
 
 /**
- * Class that encapsulates an Event task.
+ * Class that encapsulates an duke.task.Event duke.task.
  */
 public class Event extends Task {
     /**
@@ -10,7 +12,7 @@ public class Event extends Task {
     protected LocalDate timeOfEvent;
 
     /**
-     * Public constructor to create an Event task
+     * Public constructor to create an duke.task.Event duke.task
      *
      * @param description Description of the event.
      * @param timeOfEvent Time of the event.
@@ -21,9 +23,9 @@ public class Event extends Task {
     }
 
     /**
-     * String representation of an event task.
+     * String representation of an event duke.task.
      *
-     * @return String representation of an event task.
+     * @return String representation of an event duke.task.
      */
     @Override
     public String toString() {
