+++ src/main/java/duke/task/TaskList.java
@@ -1,7 +1,9 @@
+package duke.task;
+
 import java.util.ArrayList;
 
 /**
- * A task list to keep track of the tasks.
+ * A duke.task list to keep track of the tasks.
  */
 public class TaskList {
     /**
@@ -15,14 +17,14 @@ public class TaskList {
     private int length = 0;
 
     /**
-     * Empty TaskList constructor.
+     * Empty duke.task.TaskList constructor.
      */
     public TaskList() {
         this.tasks = new ArrayList<>();
     }
 
     /**
-     * TaskList constructor with tasks.
+     * duke.task.TaskList constructor with tasks.
      */
     public TaskList(ArrayList<Task> tasks) {
         this.tasks = tasks;
@@ -30,30 +32,30 @@ public class TaskList {
     }
 
     /**
-     * Method to check if a task with the corresponding index is in the task list.
+     * Method to check if a duke.task with the corresponding index is in the duke.task list.
      *
-     * @param index Index of task to be checked.
-     * @return boolean Indicate whether the task is in the task list.
+     * @param index Index of duke.task to be checked.
+     * @return boolean Indicate whether the duke.task is in the duke.task list.
      */
     public boolean isValidTaskIndex(int index) {
         return index >= 0 && index < length;
     }
 
     /**
-     * Method to return the specific task asked for.
+     * Method to return the specific duke.task asked for.
      *
-     * @param index Index of the task desired.
-     * @return The task corresponding to the index given.
+     * @param index Index of the duke.task desired.
+     * @return The duke.task corresponding to the index given.
      */
     public Task getTask(int index) {
         return tasks.get(index);
     }
 
     /**
-     * Method to add a new task to the task list.
+     * Method to add a new duke.task to the duke.task list.
      *
-     * @param newTask The task to be added.
-     * @return A new task list that contains the required tasks.
+     * @param newTask The duke.task to be added.
+     * @return A new duke.task list that contains the required tasks.
      */
     public TaskList add(Task newTask) {
         ArrayList<Task> newList = new ArrayList<>(tasks);
@@ -62,10 +64,10 @@ public class TaskList {
     }
 
     /**
-     * Method to mark a specific task as completed.
+     * Method to mark a specific duke.task as completed.
      *
-     * @param index Index of the task to be marked as completed.
-     * @return The task that was marked as completed.
+     * @param index Index of the duke.task to be marked as completed.
+     * @return The duke.task that was marked as completed.
      */
     public Task markTaskAsCompleted(int index) {
         Task task = tasks.get(index);
@@ -74,10 +76,10 @@ public class TaskList {
     }
 
     /**
-     * Method to delete task from the task list.
+     * Method to delete duke.task from the duke.task list.
      *
-     * @param index Index of the task to be deleted.
-     * @return A new task list that contains the remaining tasks.
+     * @param index Index of the duke.task to be deleted.
+     * @return A new duke.task list that contains the remaining tasks.
      */
     public TaskList deleteTask(int index) {
         ArrayList<Task> newList = new ArrayList<>(tasks);
@@ -86,21 +88,21 @@ public class TaskList {
     }
 
     /**
-     * Gets the status of the current task list.
+     * Gets the status of the current duke.task list.
      *
-     * @return String representation of the number of tasks in the task list.
+     * @return String representation of the number of tasks in the duke.task list.
      */
     public String status() {
-        String t = this.length != 1 ? "tasks" : "task";
+        String t = this.length != 1 ? "tasks" : "duke/task";
         return
                 String.format("Now you have %d %s in the list.",
                 this.length, t);
     }
 
     /**
-     * String representation of the task list.
+     * String representation of the duke.task list.
      * @return String representation of the respective tasks
-     * in the task list.
+     * in the duke.task list.
      */
     @Override
     public String toString() {
