+++ src/main/java/duke/task/ToDo.java
@@ -1,10 +1,12 @@
+package duke.task;
+
 /**
- * Class that encapsulates an ToDo task.
+ * Class that encapsulates an duke.task.ToDo duke.task.
  */
 public class ToDo extends Task {
 
     /**
-     * Public constructor to create a ToDo task
+     * Public constructor to create a duke.task.ToDo duke.task
      *
      * @param description Description of what to do.
      */
@@ -13,9 +15,9 @@ public class ToDo extends Task {
     }
 
     /**
-     * String representation of a ToDo task.
+     * String representation of a duke.task.ToDo duke.task.
      *
-     * @return String representation of a ToDo task.
+     * @return String representation of a duke.task.ToDo duke.task.
      */
     @Override
     public String toString() {
