+++ text-ui-test/EXPECTED.TXT
@@ -1,19 +1,19 @@
     +------------------------------------------------------------+
 
-    Hello! I'm Duke
+    Hello! I'm duke.Duke
     What can I do for you?
     +------------------------------------------------------------+
 
     +------------------------------------------------------------+
 
-    Got it. I've added this task:
+    Got it. I've added this duke.task:
      [T][ ] read book
-    Now you have 1 task in the list.
+    Now you have 1 duke.task in the list.
     +------------------------------------------------------------+
 
     +------------------------------------------------------------+
 
-    Got it. I've added this task:
+    Got it. I've added this duke.task:
      [D][ ] return book  (by: Sunday)
     Now you have 2 tasks in the list.
     +------------------------------------------------------------+
